server/template-recommendation-service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
scikit-learn>=1.4.0
joblib>=1.3.2
redis>=5.0.1
aiofiles>=23.2.1
numpy>=1.26.0
pytest>=7.4.0
httpx>=0.25.0
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0
pydantic>=2.0.0


